replace: v2 → blog
