--- blog/mise.toml
@@ -1,2 +1,0 @@
-[tools]
-bun = "latest"
