+++ blog/package.json
@@ -1,42 +1,38 @@
 {
-	"name": "private-blog-v4",
+	"name": "v2",
 	"private": true,
 	"type": "module",
 	"scripts": {
 		"build": "react-router build",
+		"cf-typegen": "wrangler types",
+		"deploy": "npm run build && wrangler deploy",
 		"dev": "react-router dev",
-		"start": "react-router-serve ./build/server/index.js",
-		"prepare": "panda codegen",
-		"typecheck": "react-router typegen && tsc",
-		"lint": "bunx biome check",
-		"format": "bunx biome check --write && knip",
-		"deploy": "bun run build && wrangler deploy",
-		"preview": "bun run build && wrangler dev --local"
+		"preview": "npm run build && vite preview",
+		"typecheck": "npm run cf-typegen && react-router typegen && tsc -b",
+        "lint": "npx @biomejs/biome check",
+        "format": "npx @biomejs/biome check --write",
+        "check": "npm run typecheck && npm run lint",
+        "check:fix": "npm run typecheck && npm run format"
 	},
 	"dependencies": {
-		"@ark-ui/react": "^5.4.0",
-		"@react-router/node": "^7.4.1",
-		"@react-router/serve": "^7.4.1",
 		"isbot": "^5.1.17",
-		"lucide-react": "^0.487.0",
 		"react": "^19.0.0",
 		"react-dom": "^19.0.0",
-		"react-router": "^7.4.1"
+		"react-router": "^7.5.0"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "1.9.4",
 		"@cloudflare/vite-plugin": "^1.0.0",
-		"@pandacss/dev": "^0.53.3",
-		"@park-ui/panda-preset": "^0.43.1",
-		"@react-router/dev": "^7.4.1",
-		"@types/node": "^22.13.17",
+		"@cloudflare/workers-types": "^4.20250409.0",
+		"@react-router/dev": "^7.5.0",
+		"@tailwindcss/vite": "^4.0.0",
+		"@types/node": "^20",
 		"@types/react": "^19.0.1",
 		"@types/react-dom": "^19.0.1",
-		"knip": "^5.46.5",
-		"react-router-devtools": "^1.1.0",
-		"typescript": "^5.8.2",
-		"vite": "^6.2.4",
+		"tailwindcss": "^4.0.0",
+		"typescript": "^5.7.2",
+		"vite": "^6.2.1",
 		"vite-tsconfig-paths": "^5.1.4",
-		"wrangler": "^4.7.2"
+		"wrangler": "^4.9.1"
 	}
 }
