--- blog/panda.config.ts
@@ -1,23 +1,0 @@
-import { defineConfig } from "@pandacss/dev";
-import { createPreset } from "@park-ui/panda-preset";
-import amber from "@park-ui/panda-preset/colors/amber";
-import sand from "@park-ui/panda-preset/colors/sand";
-
-export default defineConfig({
-	// Whether to use css reset
-	preflight: true,
-	presets: [
-		createPreset({ accentColor: amber, grayColor: sand, radius: "sm" }),
-	],
-	// Where to look for your css declarations
-	include: ["./app/**/{**,.client,.server}/**/*.{js,jsx,ts,tsx}"],
-	// Files to exclude
-	exclude: [],
-	// Useful for theme customization
-	theme: {
-		extend: {},
-	},
-	// The output directory for your css system
-	outdir: "styled-system",
-	jsxFramework: "react",
-});
