--- blog/postcss.config.cjs
@@ -1,5 +1,0 @@
-module.exports = {
-	plugins: {
-		"@pandacss/dev/postcss": {},
-	},
-};
