--- blog/wrangler.toml
@@ -1,15 +1,0 @@
-name = "private-blog-v4"
-compatibility_date = "2024-04-01"
-
-# Cloudflare R2の設定（画像実装時に有効化する）
-# [[r2_buckets]]
-# binding = "BLOG_IMAGES"
-# bucket_name = "blog-images"
-# preview_bucket_name = "blog-images-dev"
-
-[vars]
-APP_ENV = "production"
-
-[env.development]
-[env.development.vars]
-APP_ENV = "development"
