--- v2/.gitignore
@@ -1,13 +1,0 @@
-.DS_Store
-/node_modules/
-*.tsbuildinfo
-
-# React Router
-/.react-router/
-/build/
-
-# Cloudflare
-.mf
-.wrangler
-.dev.vars*
-
