--- v2/app/routes/home.tsx
@@ -1,20 +1,0 @@
-import type { Route } from "./+types/home";
-
-export function meta({}: Route.MetaArgs) {
-	return [
-		{ title: "New React Router App" },
-		{ name: "description", content: "Welcome to React Router!" },
-	];
-}
-
-export function loader({ context }: Route.LoaderArgs) {
-	return { message: context.cloudflare.env.VALUE_FROM_CLOUDFLARE };
-}
-
-export default function Home({ loaderData }: Route.ComponentProps) {
-	return (
-		<div>
-			<h1>Welcome to Remix</h1>
-		</div>
-	);
-}
