--- v2/package.json
@@ -1,38 +1,0 @@
-{
-	"name": "v2",
-	"private": true,
-	"type": "module",
-	"scripts": {
-		"build": "react-router build",
-		"cf-typegen": "wrangler types",
-		"deploy": "npm run build && wrangler deploy",
-		"dev": "react-router dev",
-		"preview": "npm run build && vite preview",
-		"typecheck": "npm run cf-typegen && react-router typegen && tsc -b",
-        "lint": "npx @biomejs/biome check",
-        "format": "npx @biomejs/biome check --write",
-        "check": "npm run typecheck && npm run lint",
-        "check:fix": "npm run typecheck && npm run format"
-	},
-	"dependencies": {
-		"isbot": "^5.1.17",
-		"react": "^19.0.0",
-		"react-dom": "^19.0.0",
-		"react-router": "^7.5.0"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "1.9.4",
-		"@cloudflare/vite-plugin": "^1.0.0",
-		"@cloudflare/workers-types": "^4.20250409.0",
-		"@react-router/dev": "^7.5.0",
-		"@tailwindcss/vite": "^4.0.0",
-		"@types/node": "^20",
-		"@types/react": "^19.0.1",
-		"@types/react-dom": "^19.0.1",
-		"tailwindcss": "^4.0.0",
-		"typescript": "^5.7.2",
-		"vite": "^6.2.1",
-		"vite-tsconfig-paths": "^5.1.4",
-		"wrangler": "^4.9.1"
-	}
-}
