--- v2/react-router.config.ts
@@ -1,8 +1,0 @@
-import type { Config } from "@react-router/dev/config";
-
-export default {
-  ssr: true,
-  future: {
-    unstable_viteEnvironmentApi: true,
-  },
-} satisfies Config;
